matchmakingProject/requirements.txt
asgiref==3.8.1
Django==5.1.7
Faker==37.1.0
joblib==1.3.2
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==2.2.5
scikit-learn==1.6.1
scipy==1.15.2
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.1
